Y Wires + Claw + Slowdown VAR
Since there are 10 ports in the cortex and we're allowed to use 12, we used Y-wires to join the top left with the bottom left wheel and the top right with the bottom right wheel. We also simplified the code by using slowdown, a variable that checks the slowbutton and adjusts the speed accordingly.